First Commit
Set up git via Git Bash CLI and added local changes to repo. Currently, I have created the bowling ball, linked the camera to it as a child object for tracking the ball as it moves. Created a rough BallMovement script to allow the player to aim the ball left and right on the lane. Have not set barriers as to how far left or right the player can aim yet.