Rough Draft of Physics Engine Integration
Ball now rolls upon pressing 'W' button, but only in a straight line with an applied force. Pins have attached RigidBodies allowing them to interact with the bowling ball's RigidBody and fall normally. Next Steps include directing the ball based on where the arrow is facing, reseting the pin positions along with the ball when pressing 'Space', and ensuring that the ball resets in a standstill.