Pin Reset + Camera Follow
As title states, the pins now reset to their original position along with the ball when 'Space' is pressed. Additionally, the camera now follows the ball with a slight offset behind and above it. Still working on tracking pin positions and rotations to determine score, but might take a different approach if this one doesn't work out.