Final Attempt
Was not able to calculate score from pin position / rotation changes due to inconsistencies. Attempted adding PinCollider objects on the pins, a Layer for the floor, and a Layer for the Pin Colliders and then allowing only those two layers to interact with each other in the physics system. Then, I tried both OnCollisionEnter() and OnTriggerEnter() functions for the two, but wasn't able to accomplish it. Aside from the score/UI, the rest of the physics and game operates like a bowling game would!